Update raw_info to use RubySaml::Attributes
The 0.9.0 line of ruby-saml introduced RubySaml::Attributes instead of
the hash used previously. The new class will return either single or
multiple values for the attribute requested. By setting the Attributes
to be single value compatibl we retain some of the previous behavior.

In tests, and I assume in practice, a hash is expected for :raw_info.
The enumberable interface for Attributes is used to collect all the
attributes and use the `[]` method to get the current value for each.

The upshot is that `[]` honors the value compatibility setting.